LayoutTests/accessibility/gtk/interface-table-expected.txt

This tests whether or not table-related elements implement the table interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


test1 AXRole: AXTable true
test2 AXRole: AXRow false
test3 AXRole: AXColumnHeader false
test4 AXRole: AXColumnHeader false
test5 AXRole: AXRow false
test6 AXRole: AXRowHeader false
test7 AXRole: AXCell false
test8 AXRole: AXTable true
test9 AXRole: AXRow false
test10 AXRole: AXColumnHeader false
test11 AXRole: AXColumnHeader false
test12 AXRole: AXRow false
test13 AXRole: AXRowHeader false
test14 AXRole: AXCell false
test15 AXRole: AXTreeGrid true
test16 AXRole: AXRow false
test17 AXRole: AXColumnHeader false
test18 AXRole: AXColumnHeader false
test19 AXRole: AXRow false
test20 AXRole: AXRowHeader false
test21 AXRole: AXCell false
test22 AXRole: AXTable true
test23 AXRole: AXRow false
test24 AXRole: AXColumnHeader false
test25 AXRole: AXColumnHeader false
test26 AXRole: AXRow false
test27 AXRole: AXRowHeader false
test28 AXRole: AXCell false
PASS successfullyParsed is true

TEST COMPLETE
